중간점검

+++ Hanu.ElectroLib/Constant.cs
@@ -3,12 +3,12 @@
     public static class Constant
     {
         /// <summary> [k] = N m² / C² </summary>
-        public const float Coulomb = 8.9875517923e9F;
+        public const double Coulomb = 8.9875517923e9F;
 
         /// <summary> [ε₀] = F / m </summary>
-        public const float VPermittivity = 8.8541878128e-12F;
+        public const double VPermittivity = 8.8541878128e-12F;
 
         /// <summary> [G] = N m² / kg² </summary> 
-        public const float Gravity = 6.67384e-11F;
+        public const double Gravity = 6.67384e-11F;
     }
 }
